+++ scratch.js
@@ -24,7 +24,7 @@ import plg from './plugins/dates/src/plugin.js'
 // nlp("go fast").debug()
 // nlp("i live at 123 main street").redact({ places: false }).debug()
 // nlp("Amazon and Apple compete fiercely").redact({ places: false }).debug()
-nlp("Dr. Matt G Smith lasted three seasons").redact({ places: false }).debug()
+nlp("Dr. Matt G Smith lasted three seasons").redact({}).debug()
 // console.log(doc.percentages().json())
 
 // let spec = `100 % SCAM {Value,Value,Acronym}
